Hide duplicates
With multiple source directories it is possible to have duplicates.
Consider these source directories, `a` and `b`:

    |-- a
    |   |-- bar
    |   `-- foo
    `-- b
        |-- baz
        `-- foo

The goal is to have this:

    .bar -> a/bar
    .baz -> b/baz
    .foo -> a/foo

Note the duplcate `foo` file.

We now handle this, in `lsrc` and therefore in `rcup`. We do this by
storing a `:`-separated string of destination files (e.g. `.foo`) and
`grep`ing that string before we operate on any destination.