Handle relative dotfiles directories
The `lsrc` command (and thus the other commands, too) now supports
passing a relative directory name to `-d`. Previously, this caused only
the non-host and non-tag directories to run; now all dotfiles
directories are run for all metafiles, too.

As an example, this now works correctly:

    lsrc -d dotfiles

This still works, too:

    lsrc -d /home/mike/dotfiles