Add a release target
The `make release` command will build the Debian package, push the
version to a git tag on GitHub, and upload the docs to GitHub pages.

This also adds a `deb` target, along with `build-docs`, `upload-docs`,
`build-tag`, and `push-tag`.

In addition, introduce a `NEWS.md` file.

Both `rcm.sh.in` and `NEWS.md.in` will act as input files. This is to
abstract over the version number.