Introduce exclusion patterns
The lsrc(1), rcup(1), and rcdn(1) commands now take any number of `-e`
flags, used to specify an exclusion pattern. This can also be controlled
via rcrc(5), the `EXCLUDES` variable.

An exclusion pattern specifies a file glob to skip. In the case of
lsrc(1), any file matching the glob is not listed; in rcup(1) it is not
symlinked; and in rcdn(1) it is not removed.

The file glob can be preceded by the name of a dotfiles directory
(separated from the file glob by a colon) to increase the specificity.

Useful for:

    rcdn -e rcrc
    rcup -d work-dotfiles -e bashrc
    rcup -d ~/.dotfiles -d wife-dotfiles -d sys-dotfiles -e wife-dotfiles:tigrc