Add the COPY_ALWAYS option
The suite now honors the `COPY_ALWAYS` option in rcrc(5). This can be
set to a space-separated list of file globs. Any file matching a glob is
copied instead of symlinked. This is handy both for secure programs
(`netrc`, `ssh/id_*`) and for programs that oddly re-write files
(`weechat/*`).

To always copy everything, use the `*` glob.

This is reflected throughout the suite as follows:

* lsrc now has a `-F` option which shows a symbol to indicate whether it
  is a symlink (`@`) or a copy (`X`).
* rcdn only removes symlinks unless the file under question matches a
  `COPY_ALWAYS` glob, in which case it is removed regardless of whether
  it is a symlink.
* rcup will copy instead of symlinking any file that matches any
  `COPY_ALWAYS` glob.