Add RCM_CONFIG environment variable
Providing an alternate config as a command-line flag would be impossible
due to the required timing (it must be sourced before option parsing for
most of rcm to work correctly).

So I chose to use the RCM_CONFIG environment variable. I also moved the
sourcing to occur at the end of loading rcm.sh always.