Show a difference between moving and link in mkrc
I found that after you add a file with `mkrc` command from any dir, the output
looks like:

    $ mkrc ~/.my-awesome-dot-file
    ‘$HOME/.my-awesome-dot-file -> ‘$HOME/.dotfiles/my-awesome-dot-file'
    ‘$HOME/.my-awesome-dot-file -> ‘$HOME/.dotfiles/my-awesome-dot-file'
    $

It looks like the output it's appearing twice, which could led you believe
something went wrong or the output is a bit buggy. Obivously, it's not, and the
command ran without problems.

Use `$PRINT` to show which step it is on before the verbose messages are
displayed.