Add better feedback to the user when adding new files
I found that after you add a file with `mkrc` command from any dir, the output
looks like:

    $ mkrc ~/.my-awesome-dot-file
    ‘$HOME/.my-awesome-dot-file -> ‘$HOME/.dotfiles/my-awesome-dot-file'
    ‘$HOME/.my-awesome-dot-file -> ‘$HOME/.dotfiles/my-awesome-dot-file'
    $

It looks like the output it's appearing twice, which could led you believe
something went wrong or the output is a bit buggy. Obivously, it's not, and the
command ran without problems. This little patch improves the feedback to the
user making clear that those are two different process rather than one that's
being repeated.