Unset the CDPATH before running the program
As reported in #23, you can trick `rcup` into installing into `.` with
this `CDPATH`:

    export CDPATH=/tmp:.

As described [online][1], it is a mistake to export the `CDPATH`
environment variable to begin with. Since it is a one-liner to work
around it, though, it's worth fixing.

The workaround is to unset `CDPATH` at the beginning of the script. This
does not affect the outside environment, it only affects the script and
its functions.

[1]: http://bosker.wordpress.com/2012/02/12/bash-scripters-beware-of-the-cdpath/