Add a list of contributors to rcm(7)
This list is an alphabetical listing of everyone who has authored at
least one commit, plus their email address. When the author is a
coworker I have changed this to use their thoughtbot email address.

This also updates the `AUTHORS` section to mention thoughtbot.

The `NEWS.md.in` file has been changed to reflect that 1.2.0 is
released.

A `CONTRIBUTING.md` file has been added to explain what is required to
send a good contribution.