Add a mailmap
This is documented in Documentation/mailmap.txt for git, which can be
seen online: https://github.com/git/git/blob/6a907786af835ac15962be53f1492f23e044f479/Documentation/mailmap.txt

This is used by git-shortlog and git-check-mailmap.

The big picture in this branch is to move all the contributor listings
to this one file, then generate rcm.7 from that.