mkrc with relative filenames inside dotted dir
This is best explained with an example. If I want to track a file like
`~/.bundle/config`, the correct way would be:

    ~$ mkrc ~/.bundle/config
    --> ~/.dotfiles/bundle/config

But if you are already inside the directory, say:

    ~/.bundle $

and you ran:

    ~/.bundle $ mkrc con<TAB> # for autocomplete
    ~/.bundle $ mkrc config
    --> ~/.dotfiles/.config

Which is obviously not what you meant.

This basically checks first if the file exists in the current working
directory and if it is, it's expand the full path.

    ~/.bundle $ mkrc config
    --> ~/.dotfiles/bundle/config