Add a hostname override
Based on issue #82, we now provide `-B` to override the hostname. In
particular:

- `mkrc -B foo` will enable `-o` but with the hostname set up `foo`.
- `lsrc -B foo` will work like normal `lsrc` except it treats `host-foo`
as the host-specific directory.
- `rcup -B foo` will run a normal `rcup` except `host-foo` is the
host-specific directory.
- `rcdn -B foo` is just like normal `rcdn`, but with `host-foo` as the
  host-specific directory.

The `HOSTNAME` can also be set in the rcrc(5), and this is overridden by
the aforementioned `-B`.

While making this change: The `test/Makefile.am` used a mix of tabs and
spaces. Since it's a Makefile, replace it all with tabs.