Update documentation on usage
New flags have accumulated without proper care for the usage
instructions or man pages.

I manually went through each program and verified its usage instruction
against its `getopts`, then I alphabeticalized the usage message.

Based on the usage message, I then verified the synposis in the manpage.
Then, based on the synposis, I alphabeticalized the detailed listing of
the arguments and filled in the missing pieces. The `-h` and `-V`
arguments were missing from all manpages.

In the future we will need to be more careful about this. It would be
good to automate a checker that refuses to build unless the docs have
all the flags mentioned.