Fix hooks so that they are run in the directory they are located
Calls `find` with the `-execdir` action instead of `-exec`, or if `-execdir`
is unavailable, uses `-exec` with a shell one-liner to simulate the former.
These changes allow hooks to refer to dotfiles with relative paths. For
instance, we can call a Makefile two directories up simply with `make -C ../..`.
Also the use of `-execdir` is supposed to be more secure, where available.

In order to achieve the above:
- add support for local (aclocal) m4 macros
- add a test to configure to search for a find(1) with the `-execdir` capability
- set up a substitution variable (FIND_HAS_EXECDIR) and a file substitution
  (code snippet) accordingly