Fix hooks so that they are run in the directory where they are located
Calls `find(1)` with the `-exec` action just like before but instead of `run_hooks()`
letting `find` execute the hook directly it is wrapped in a shell one-liner which
changes to the hook's directory and executes the hook with "./" prepended to its
basename. These changes allow hooks to refer to dotfiles with relative paths.
For instance we can call a Makefile two directories up simply with `make -C ../..`.

Also make sure we are compatible with Solaris' Bourne shell and `find(1)`.

Closes #150. Closes #149.