Handle spaces in dotfile name
In mkrc, separate the list of files with newlines instead of spaces.
Change the `$IFS` when iterating to handle this.

We hand the file off to rcup, which encodes the file name by replacing
spaces with the bell character (`\a`).

rcup then sends the file name off to lsrc, which decodes the bell back
into a space.

The test makes sure an `a` character is in the filename, in case some
encoding goes wrong. We use tr(1) instead of sed(1) because tr(1)
handles `\a`.

Shoutout to Sublime Text 3 for forcing this issue.