Update Ubuntu installation instructions
As of Ubuntu 19.04 (Disco Dingo), rcm is now available in the Universe
package repository.

The ppa will no longer be necessary for future versions of Ubuntu but
will be kept around until the older releases are EOL.

Also, add a line to ensure that `software-properties-common` is
installed. It is for instance not included in the bare `ubuntu:16.04`
docker image.

Add software-properties-common