Fewer sharp edges in mkrc
Update the documentation for mkrc(1) to clarify some common gotchas:

- Add the new file normally, then use mkrc(1) to do the dotfiles dance.
- The `-C` flag copies files and, as such, this means they are no longer
tracked. We can manage symlinks for you, but the `-C` flag is a
convenience option that takes you outside of the area we can manage.
- We really mean it about `-C` when it comes to directories. Copy it all
you want, but it is now your responsibility.

Fixes #206.

Another sharp edge is around the dotfiles directory.

- Only support one `-d` flag. It doesn't make sense to pass multiple
`-d` flags to mkrc(1).
- The argument of the `-d` flag must be a directory.

Fixes #254.