Update tests for new flag
In 4fcb2b, the `-k` flag was introduced.  It seems that due to an
oversight, the tests weren't updated to take this flag into account and
are currently failing on a fresh clone of the project.  This commit
updates the `test/mkrc-one-dotfile-arg.t` test to account for the new
flag.

Fixes #278