Clarify the SHA256 installation programs
Most people will be manually installing this using a GNU userland, so
show `sha256sum` as the default, but highlight that it might be spelled
differently for different OSes.

When GNU [added it to coreutils in 2005][], they called it `sha256sum`, to
match their existing `sha1sum`. They [renamed][] `shasum` to `sha1sum` in
2000, [called it][] `shasum` earlier in 2000 by analogy to `md5sum`, and
[named it][] `md5sum` in 1995.

OpenBSD [added the program][] in 2004, extending the existing `md5` program to
support being called as `sha256`.

It looks like Illumos and Solaris [ship with][] GNU coreutils.

RFC 6234 calls the SHA256 sample program `shatest` in section 8.5. RFC
1321 calls the sample MD5 program `mddriver` in Appendix A.

[ship with]: https://docs.oracle.com/cd/E88353_01/html/E37839/sha256sum-1.html
[added the program]: https://cvsweb.openbsd.org/src/bin/md5/md5.c?rev=1.29&content-type=text/x-cvsweb-markup
[added it to coreutils in 2005]: https://git.savannah.gnu.org/cgit/coreutils.git/commit/?id=0ebfe3cf4b83f016ab2b6c3dc6126e9e8621e0a9
[renamed]: https://git.savannah.gnu.org/cgit/coreutils.git/commit/?id=7bb9c7382ff3d4a31aa6e79bd03b2890d5d0ffe6
[called it]: https://git.savannah.gnu.org/cgit/coreutils.git/commit/?id=148e5d1b4d2d9ab1a5f9523e2e1f92914f0f2be0
[named it]: https://git.savannah.gnu.org/cgit/coreutils.git/commit/?id=84876e7e9e2b42a430aff37adc0888c4562b447a